refactor: split central test runner responsibilities
